// src/types/Harem++.ts
export type HaremPlusPlusAllGirlsCache_Incomplete = {
  id: string;
  icon0: string;
  poseImage0: string;
  shards: number;
  name: string;
  rarity: number; // 0,1,2,3,4,5
  gradeSkins: Array<{
    id_girl_grade_skin: number;
    grade_skin_name: string;
    ico_path: string;
    image_path: string;
    num_order: number;
  }>;
};